Added 5 GHz config

+++ setup.sh
@@ -3,7 +3,14 @@ COUNTRY=DE
 INTERNAL_WLAN=aa:bb:cc:dd:ee:ff
 EXTERNAL_WLAN=aa:bb:cc:dd:ee:ee
 RANGE_WLAN=1
+
+# 2.4 GHz
+MODE=g
 CHANNEL=7
+
+# 5 GHz
+# MODE=a
+# CHANNEL=36
 
 INTERNAL_ETH=11:22:33:44:55:66
 EXTERNAL_ETH=11:22:33:44:66:66
@@ -20,6 +27,8 @@ orig_file() {
     sudo cp $1.orig $1
 }
 
+echo $COUNTRY | sudo tee /country.txt
+
 # Fix interfaces
 echo "SUBSYSTEM==\"net\", ACTION==\"add\", DRIVERS==\"?*\", ATTR{address}==\"$EXTERNAL_WLAN\", ATTR{type}==\"1\", NAME=\"wlan0\"
 SUBSYSTEM==\"net\", ACTION==\"add\", DRIVERS==\"?*\", ATTR{address}==\"$INTERNAL_WLAN\", ATTR{type}==\"1\", NAME=\"wlan1\"
@@ -35,7 +44,7 @@ static ip_address=192.168.$RANGE_ETH.1/24" | sudo tee -a /etc/dhcpcd.conf
 
 # Create DHCP config
 echo "dhcp-range=wlan1,192.168.$RANGE_WLAN.2,192.168.$RANGE_WLAN.255,255.255.255.0,24h
-dhcp-range=wlan1,192.168.$RANGE_ETH.2,192.168.$RANGE_ETH.255,255.255.255.0,24h" | sudo tee /etc/dnsmasq.conf
+dhcp-range=eth1,192.168.$RANGE_ETH.2,192.168.$RANGE_ETH.255,255.255.255.0,24h" | sudo tee /etc/dnsmasq.conf
 
 # Get WiFi settings
 echo -n "SSID: "
@@ -47,7 +56,7 @@ read PASSWORD
 echo "interface=wlan1
 ssid=$SSID
 channel=$CHANNEL
-hw_mode=g
+hw_mode=$MODE
 ieee80211n=1
 ieee80211d=1
 country_code=$COUNTRY
@@ -68,11 +77,16 @@ chmod +x leases.sh mac.sh net_led.sh rtl8192eu.sh
 orig_file /etc/sysctl.conf
 echo "net.ipv4.ip_forward=1" | sudo tee -a /etc/sysctl.conf
 
-orig_file /etc/rc.local
-sudo iptables -t nat -A POSTROUTING -o wlan0 -j MASQUERADE
-sudo iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
+if [ ! -f /etc/iptables.ipv4.nat ]; then
+    sudo iptables -t nat -A POSTROUTING -o wlan0 -j MASQUERADE
+    sudo iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
 
-sudo iptables-save | sudo tee /etc/iptables.ipv4.nat
+    sudo iptables-save | sudo tee /etc/iptables.ipv4.nat
+fi
 
 orig_file /etc/rc.local
-sudo sed -i '/^exit 0/i iptables-restore </etc/iptables.ipv4.nat\n/home/pi/net_led.sh &' /etc/rc.local
+sudo sed -i '/^exit 0/i iptables-restore </etc/iptables.ipv4.nat\n/home/pi/mac.sh wlan0 $(cat /mac.txt)\n/home/pi/net_led.sh &' /etc/rc.local
+
+sudo mv /etc/wpa_supplicant/wpa_supplicant.conf /etc/wpa_supplicant/wpa_supplicant-wlan0.conf
+
+sudo reboot
